+++ src/main/java/com/jiwon/blog/dto/request/PostRequest.java
@@ -1,4 +1,4 @@
-package com.jiwon.blog.dto;
+package com.jiwon.blog.dto.request;
 
 import com.jiwon.blog.entity.Category;
 import com.jiwon.blog.entity.Member;
