+++ src/main/java/com/jiwon/blog/service/CategoryService.java
@@ -1,7 +1,7 @@
 package com.jiwon.blog.service;
 
-import com.jiwon.blog.dto.CategoryRequest;
-import com.jiwon.blog.dto.CategoryResponse;
+import com.jiwon.blog.dto.request.CategoryRequest;
+import com.jiwon.blog.dto.response.CategoryResponse;
 import com.jiwon.blog.entity.Category;
 import com.jiwon.blog.repository.CategoryRepository;
 import lombok.extern.slf4j.Slf4j;
